requirements.txt
pandas
numpy
streamlit
folium
streamlit-folium
inflection
Pillow
plotly